update models with reformatting, add more env vars
