+++ workspace/marugoto_mri/model/main.py
@@ -20,7 +20,15 @@ from mil.helpers import (
 
 scratchDir = os.getenv('SCRATCH_DIR', '/platform/scratch')
 dataDir = os.getenv('DATA_DIR', '/platform/data/')
+num_epochs = os.getenv('MAX_EPOCHS', 64)
+min_peers = os.getenv('MIN_PEERS', 2)
+max_peers = os.getenv('MAX_PEERS', 7)
+local_compare_flag = os.getenv('LOCAL_COMPARE_FLAG', False)
+useAdaptiveSync = os.getenv('USE_ADAPTIVE_SYNC', False)
+syncFrequency = os.getenv('SYNC_FREQUENCY', 32)
+
 current_time = datetime.now().strftime("%Y_%m_%d_%H%M%S")
+
 
 data_split = '40-30-10-20'
 #data_split = '25-25-25-25'
@@ -35,7 +43,14 @@ if __name__ == "__main__":
     slide_csv = Path(os.path.join(dataDir, 'slide_table.csv')),
     feature_dir = Path(feature_dir_path),
     output_path = Path(out_dir),
-    target_label = "Malign")
+    target_label = "Malign",
+    n_epoch = num_epochs,
+    local_compare_flag = local_compare_flag,
+    min_peers = min_peers,
+    max_peers = max_peers,
+    useAdaptiveSync = useAdaptiveSync,
+    syncFrequency = syncFrequency,
+    )
 
     deploy_categorical_model_(
     clini_table = Path(os.path.join(dataDir, 'clinical_table.csv')),
