+++ workspace/marugoto_mri/model/main_local_compare(deprecated).py
