+++ workspace/marugoto_mri/swci/swci-init_ori
@@ -34,25 +34,7 @@ create task from taskdefs/swarm_task.yaml
 finalize task swarm_task_<TIME_STAMP>
 get task info swarm_task_<TIME_STAMP>
 get task body swarm_task_<TIME_STAMP>
-list tasks
-EXIT ON FAILURE OFF
 
-# Assign run task
-EXIT ON FAILURE
-RESET TASKRUNNER defaulttaskbb.taskdb.sml.hpe
-ASSIGN TASK swarm_task_<TIME_STAMP> TO defaulttaskbb.taskdb.sml.hpe WITH <NUM-MIN_PEERS> PEERS
-WAIT FOR TASKRUNNER defaulttaskbb.taskdb.sml.hpe
-# Grace period for ML container to save the model,
-# before the below reset command cleans up the ML and SL Containers.
-SLEEP 15
-RESET TASKRUNNER defaulttaskbb.taskdb.sml.hpe
-EXIT ON FAILURE OFF
-
-
-# Build task was already run. Now build and run swarm run tasks
-
-# Create and finalize swarm run task
-EXIT ON FAILURE
 create task from taskdefs/swarm_task_local_compare.yaml
 finalize task swarm_task_local_compare_<TIME_STAMP>
 get task info swarm_task_local_compare_<TIME_STAMP>
@@ -63,6 +45,7 @@ EXIT ON FAILURE OFF
 # Assign run task
 EXIT ON FAILURE
 RESET TASKRUNNER defaulttaskbb.taskdb.sml.hpe
+ASSIGN TASK swarm_task_<TIME_STAMP> TO defaulttaskbb.taskdb.sml.hpe WITH <NUM-MIN_PEERS> PEERS
 ASSIGN TASK swarm_task_local_compare_<TIME_STAMP> TO defaulttaskbb.taskdb.sml.hpe WITH <NUM-MIN_PEERS> PEERS
 WAIT FOR TASKRUNNER defaulttaskbb.taskdb.sml.hpe
 # Grace period for ML container to save the model,
