+++ workspace/marugoto_mri/swci/taskdefs/swarm_task_local_compare_ori.yaml
@@ -4,10 +4,18 @@ Author: Jeff
 Prereq: user_env_build_task_<TIME_STAMP>
 Outcome: swarm_task_local_compare
 Body:
-  Command: model/main_local_compare.py
+  Command: model/main.py
   Entrypoint: python3
   WorkingDir: /tmp/test
-  Envvars: [ "DATA_DIR": data-and-scratch/data, "SCRATCH_DIR": data-and-scratch/scratch, "MODEL_DIR": model, "MAX_EPOCHS": <NUM-MAX_EPOCHS>, "MIN_PEERS": <NUM-MIN_PEERS> ]
+  Envvars: [ "DATA_DIR": data-and-scratch/data,
+             "SCRATCH_DIR": data-and-scratch/scratch,
+             "MODEL_DIR": model,
+             "MAX_EPOCHS": <NUM-MAX_EPOCHS>,
+             "MIN_PEERS": <NUM-MIN_PEERS>,
+             "LOCAL_COMPARE_FLAG": True,
+             "USE_ADAPTIVE_SYNC": True,
+             "SYNC_FREQUENCY": 32,
+           ]
   PrivateContent: /tmp/test/data-and-scratch
   SharedContent:
     - Src: /opt/hpe/swarm-learning-hpe/workspace/<MODULE-NAME>/model
