+++ workspace/marugoto_mri/swci/taskdefs/user_env_build_task_ori.yaml
@@ -1,4 +1,43 @@
-Name: user_env_build_task_<TIME_STAMP>
+Name: user_env_build_task_<TIME_STAMP># Create and finalize swarm run task
+EXIT ON FAILURE
+create task from taskdefs/swarm_task.yaml
+finalize task swarm_task_<TIME_STAMP>
+get task info swarm_task_<TIME_STAMP>
+get task body swarm_task_<TIME_STAMP>
+list tasks
+EXIT ON FAILURE OFF
+
+# Assign run task
+EXIT ON FAILURE
+RESET TASKRUNNER defaulttaskbb.taskdb.sml.hpe
+ASSIGN TASK swarm_task_<TIME_STAMP> TO defaulttaskbb.taskdb.sml.hpe WITH <NUM-MIN_PEERS> PEERS
+WAIT FOR TASKRUNNER defaulttaskbb.taskdb.sml.hpe
+# Grace period for ML container to save the model,
+# before the below reset command cleans up the ML and SL Containers.
+SLEEP 15
+RESET TASKRUNNER defaulttaskbb.taskdb.sml.hpe
+EXIT ON FAILURE OFF
+
+# Assign run compare task
+EXIT ON FAILURE
+RESET TASKRUNNER defaulttaskbb.taskdb.sml.hpe
+ASSIGN TASK swarm_task_local_compare_<TIME_STAMP> TO defaulttaskbb.taskdb.sml.hpe WITH <NUM-MIN_PEERS> PEERS
+WAIT FOR TASKRUNNER defaulttaskbb.taskdb.sml.hpe
+# Grace period for ML container to save the model,
+# before the below reset command cleans up the ML and SL Containers.
+SLEEP 15
+RESET TASKRUNNER defaulttaskbb.taskdb.sml.hpe
+EXIT ON FAILURE OFF
+
+# List and reset training contract
+EXIT ON FAILURE
+LIST CONTRACTS
+RESET CONTRACT defaultbb.cqdb.sml.hpe
+EXIT ON FAILURE OFF
+
+# Exit
+EXIT
+
 TaskType: MAKE_USER_CONTAINER
 Author: Jeff
 Prereq: ROOTTASK
