+++ workspace/odelia-breast-mri/model/main_local_compare(deprecated).py
