+++ workspace/odelia-breast-mri/model/models/__init__.py
@@ -1,1 +1,2 @@
 from .base_model import VeryBasicModel, BasicModel, BasicClassifier
+from .resnet import ResNet
