+++ workspace/odelia-breast-mri/swci/taskdefs/swarm_task_ori.yaml
@@ -7,7 +7,15 @@ Body:
   Command: model/main.py
   Entrypoint: python3
   WorkingDir: /tmp/test
-  Envvars: [ "DATA_DIR": data-and-scratch/data, "SCRATCH_DIR": data-and-scratch/scratch, "MODEL_DIR": model, "MAX_EPOCHS": <NUM-MAX_EPOCHS>, "MIN_PEERS": <NUM-MIN_PEERS> ]
+  Envvars: [ "DATA_DIR": data-and-scratch/data,
+             "SCRATCH_DIR": data-and-scratch/scratch,
+             "MODEL_DIR": model,
+             "MAX_EPOCHS": <NUM-MAX_EPOCHS>,
+             "MIN_PEERS": <NUM-MIN_PEERS>,
+             "LOCAL_COMPARE_FLAG": False,
+             "USE_ADAPTIVE_SYNC": True,
+             "SYNC_FREQUENCY": 512,
+           ]
   PrivateContent: /tmp/test/data-and-scratch
   SharedContent:
     - Src: /opt/hpe/swarm-learning-hpe/workspace/<MODULE-NAME>/model
